20.1.1

+++ Application/include/nx/PlayData.hpp
@@ -24,19 +24,19 @@ namespace NX {
 
     // PlayEvents are parsed PdmPlayEvents containing only necessary information
     struct PlayEvent {
-        PlayEventType type;        // Type of PlayEvent
-        AccountUid userID;         // UserID
-        TitleID titleID;           // TitleID
-        EventType eventType;       // See EventType enum
-        u64 clockTimestamp;        // Time of event
-        u64 steadyTimestamp;       // Steady timestamp (used for calculating duration)
+        PlayEventType type;     // Type of PlayEvent
+        AccountUid userID;      // UserID
+        TitleID titleID;        // TitleID
+        EventType eventType;    // See EventType enum
+        u64 clockTimestamp;     // Time of event
+        u64 steadyTimestamp;    // Steady timestamp (used for calculating duration)
     };
 
     // A PlaySession represents a session of play for a game. It contains the start
     // and end timestamps and playtime for convenience. Note that (end-start) != playtime
     // due to time when the game may not have been focussed.
     struct PlaySession {
-        u64 playtime;           // Total playtime in seconds
+        u32 playtime;           // Total playtime in seconds
         u64 startTimestamp;     // Time of launch
         u64 endTimestamp;       // Time of exit
     };
@@ -44,9 +44,9 @@ namespace NX {
     // PdmPlayStatistics but only the necessary things
     struct PlayStatistics {
         TitleID titleID;        // TitleID of these stats
-        u64 firstPlayed;        // Timestamp of first launch
-        u64 lastPlayed;         // Timestamp of last play (exit)
-        u64 playtime;           // Total playtime in seconds
+        u32 firstPlayed;        // Timestamp of first launch
+        u32 lastPlayed;         // Timestamp of last play (exit)
+        u32 playtime;           // Total playtime in seconds
         u32 launches;           // Total launches
     };
 
@@ -54,7 +54,7 @@ namespace NX {
     // only contains recent values
     struct RecentPlayStatistics {
         TitleID titleID;        // TitleID of these statistics
-        u64 playtime;           // Total playtime in seconds
+        u32 playtime;           // Total playtime in seconds
         u32 launches;           // Total launches
     };
 
@@ -118,10 +118,10 @@ namespace NX {
             // Returns all play sessions for the given title ID and user ID
             std::vector<PlaySession> getPlaySessionsForUser(TitleID, AccountUid);
 
-            // Returns a RecentPlayStatistics for the given time range and user ID for all Title IDs
+            // Returns a RecentPlayStatistics for the given time range for all users
             RecentPlayStatistics * getRecentStatisticsForUser(u64, u64, AccountUid);
 
-            // Returns a RecentPlayStatistics for the given time range, user ID and title ID
+            // Returns a RecentPlayStatistics for the given time range and user ID
             RecentPlayStatistics * getRecentStatisticsForTitleAndUser(TitleID, u64, u64, AccountUid);
 
             // Returns a PlayStatistics for the given titleID and userID
